requirements/runtime.txt
matplotlib
numpy
pycocotools
scipy
six
terminaltables
timm
einops
transformers
omegaconf
pytorch_lightning
